requirements.txt
astroid==2.3.3
certifi==2020.12.5
colorama==0.4.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pylint==2.4.4
pytesseract==0.3.7
typed-ast==1.4.0
wincertstore==0.2
